Update qst.py
Added new mode, WXGPSA. This mode read a WX file with the information from a weather station. the file format must be like APRS file. with two lines like this.

2021-02-18T08:00:22
170/006g006t057r000p000P000h48b10110

this mod add more leght to GPS-A message to support the size of weather text. also is added more times in the sellection times.