Chat bytearray fix
The conversion of the transport layer to use bytearray on python 2
to be compatible with both python 3 and newer python is still having
unattended consquences.

d_rats/sessions/chat.py(_incoming_chat, incoming_data):

Properly convert the framedata from bytearray to Unicode with
replacement done for invalid characters.

d_rats/sessions/chat.py(incoming_data):

In a bytearray, the array elements are already integers, so
a conversion of the status code from ASCII is required.

Raising an generic exception invalid input from a remote system allows
a denial of service attack from a remote system, so for now just log
the issue to the console.  Eventually this should be logged to the event
log.