config gps_default_message fix
Remove the "*20" from the gps default message.
The asterisk character is used by DPRS to start the checksum so its
presence here causes d-rats to flood the console with mesages since
the checksum does not match.

Not sure if the trailing spaces are still needed, that can be handled
in a later commit.